Fixing up the save of the Many-to-Many
According to the Hibernate guys, this is actually the correct way to do things. I'm not really sure where things were blowing up, but this fixes it.

The collection types really should be sets, if you think about them as a tag only belonging to the same post once, and a post isn't going to have the same tag more than once either.

Like I said earlier, I really don't know why this was broken, but this does fix it. It doesn't make 100% sense to me though.